+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### 🎨 Enhancements
 
+- [#293](https://github.com/estruyf/vscode-front-matter/issues/293): Support added for setting preview images in block fields
 - [#294](https://github.com/estruyf/vscode-front-matter/issues/294): Full-text search allows you to search through all your page content
 
 ### ⚡️ Optimizations